feat: Restructure settings and add mirroring
